refactor(api): migrate GET /api/campaigns/{id}/collective-results to prescription Bounded Context
