+++ api/src/prescription/campaign/application/campaign-results-controller.js
@@ -1,7 +1,9 @@
+import { extractLocaleFromRequest } from '../../../../lib/infrastructure/utils/request-response-utils.js';
 import { extractParameters } from '../../../shared/infrastructure/utils/query-params-utils.js';
 import { certificabilityByLabel } from '../../shared/application/helpers.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as campaignAssessmentResultMinimalSerializer from '../infrastructure/serializers/jsonapi/campaign-assessment-result-minimal-serializer.js';
+import * as campaignCollectiveResultSerializer from '../infrastructure/serializers/jsonapi/campaign-collective-result-serializer.js';
 import * as campaignProfilesCollectionParticipationSummarySerializer from '../infrastructure/serializers/jsonapi/campaign-profiles-collection-participation-summary-serializer.js';
 
 const findAssessmentParticipationResults = async function (request) {
@@ -47,9 +49,19 @@ const findProfilesCollectionParticipations = async function (request) {
   return campaignProfilesCollectionParticipationSummarySerializer.serialize(results);
 };
 
+const getCollectiveResult = async function (request, h, dependencies = { campaignCollectiveResultSerializer }) {
+  const { userId } = request.auth.credentials;
+  const campaignId = request.params.id;
+  const locale = extractLocaleFromRequest(request);
+
+  const campaignCollectiveResult = await usecases.computeCampaignCollectiveResult({ userId, campaignId, locale });
+  return dependencies.campaignCollectiveResultSerializer.serialize(campaignCollectiveResult);
+};
+
 const campaignResultsController = {
   findAssessmentParticipationResults,
   findProfilesCollectionParticipations,
+  getCollectiveResult,
 };
 
 export { campaignResultsController };
