+++ api/src/prescription/campaign/domain/read-models/CampaignCollectiveResult.js
