+++ api/src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-collective-result-serializer.js
