+++ api/tests/acceptance/application/campaigns/campaign-controller_test.js
@@ -17,146 +17,6 @@ describe('Acceptance | API | Campaign Controller', function () {
     server = await createServer();
   });
 
-  describe('GET /api/campaigns/{id}/collective-result', function () {
-    const assessmentStartDate = '2018-01-02';
-    const participationStartDate = '2018-01-01';
-
-    let userId;
-
-    beforeEach(async function () {
-      userId = databaseBuilder.factory.buildUser({ firstName: 'Jean', lastName: 'Bono' }).id;
-      organization = databaseBuilder.factory.buildOrganization();
-      campaign = databaseBuilder.factory.buildCampaign({
-        name: 'Campagne de Test N°2',
-        organizationId: organization.id,
-        idPixLabel: 'Identifiant entreprise',
-      });
-      databaseBuilder.factory.buildCampaignSkill({ campaignId: campaign.id, skillId: 'recSkillId1' });
-
-      databaseBuilder.factory.buildMembership({
-        userId,
-        organizationId: organization.id,
-        organizationRole: Membership.roles.MEMBER,
-      });
-
-      campaign = databaseBuilder.factory.buildCampaign({
-        name: 'Campagne de Test N°3',
-        organizationId: organization.id,
-      });
-      databaseBuilder.factory.buildCampaignSkill({ campaignId: campaign.id, skillId: 'recSkillId1' });
-      databaseBuilder.factory.buildCampaignSkill({ campaignId: campaign.id, skillId: 'recSkillId2' });
-
-      const campaignParticipation = databaseBuilder.factory.buildCampaignParticipation({
-        campaignId: campaign.id,
-        userId,
-        createdAt: new Date(participationStartDate),
-        sharedAt: new Date('2018-01-27'),
-      });
-
-      const assessment = databaseBuilder.factory.buildAssessment({
-        userId,
-        type: 'CAMPAIGN',
-        createdAt: new Date(assessmentStartDate),
-        campaignParticipationId: campaignParticipation.id,
-      });
-
-      databaseBuilder.factory.buildKnowledgeElement({
-        skillId: 'recSkillId1',
-        status: 'validated',
-        userId,
-        assessmentId: assessment.id,
-        competenceId: 'recCompetence1',
-        createdAt: new Date('2017-12-01'),
-      });
-
-      await databaseBuilder.commit();
-
-      const learningContent = [
-        {
-          id: 'recArea1',
-          title_i18n: {
-            fr: 'area1_Title',
-          },
-          color: 'specialColor',
-          competences: [
-            {
-              id: 'recCompetence1',
-              name_i18n: { fr: 'Fabriquer un meuble' },
-              index: '1.1',
-              tubes: [
-                {
-                  id: 'recTube1',
-                  skills: [
-                    {
-                      id: 'recSkillId1',
-                      nom: '@web2',
-                      challenges: [],
-                    },
-                    {
-                      id: 'recSkillId2',
-                      nom: '@web3',
-                      challenges: [],
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-        },
-      ];
-      const learningContentObjects = learningContentBuilder.fromAreas(learningContent);
-      mockLearningContent(learningContentObjects);
-    });
-
-    it('should return campaign collective result with status code 200', async function () {
-      // given
-      const url = `/api/campaigns/${campaign.id}/collective-results`;
-      const request = {
-        method: 'GET',
-        url,
-        headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-      };
-      const expectedResult = {
-        data: {
-          type: 'campaign-collective-results',
-          id: campaign.id.toString(),
-          attributes: {},
-          relationships: {
-            'campaign-competence-collective-results': {
-              data: [
-                {
-                  id: `${campaign.id}_recCompetence1`,
-                  type: 'campaignCompetenceCollectiveResults',
-                },
-              ],
-            },
-          },
-        },
-        included: [
-          {
-            id: `${campaign.id}_recCompetence1`,
-            type: 'campaignCompetenceCollectiveResults',
-            attributes: {
-              'area-code': '1',
-              'area-color': 'specialColor',
-              'average-validated-skills': 1,
-              'competence-id': 'recCompetence1',
-              'competence-name': 'Fabriquer un meuble',
-              'targeted-skills-count': 2,
-            },
-          },
-        ],
-      };
-
-      // when
-      const response = await server.inject(request);
-
-      // then
-      expect(response.statusCode).to.equal(200, response.payload);
-      expect(response.result).to.deep.equal(expectedResult);
-    });
-  });
-
   describe('GET /api/campaigns/{id}/analyses', function () {
     let userId;
