+++ api/tests/prescription/campaign/unit/application/campaign-results-route_test.js
@@ -201,4 +201,33 @@ describe('Unit | Application | campaign-results-router ', function () {
       expect(result.statusCode).to.equal(400);
     });
   });
+
+  describe('GET /api/campaigns/{id}/collective-results', function () {
+    it('should return 200', async function () {
+      // given
+      sinon
+        .stub(campaignResultsController, 'getCollectiveResult')
+        .callsFake((request, h) => h.response('ok').code(200));
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/campaigns/1/collective-results');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('GET', '/api/campaigns/invalid/collective-results');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
 });
