+++ api/tests/prescription/campaign/unit/infrastrucutre/serializers/jsonapi/campaign-collective-result-serializer_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/campaign-collective-result-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../../src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-collective-result-serializer.js';
+import { domainBuilder, expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | campaign-collective-results-serializer', function () {
   describe('#serialize', function () {
