+++ api/tests/tooling/domain-builder/factory/build-campaign-collective-result.js
@@ -1,4 +1,4 @@
-import { CampaignCollectiveResult } from '../../../../lib/domain/read-models/CampaignCollectiveResult.js';
+import { CampaignCollectiveResult } from '../../../../src/prescription/campaign/domain/read-models/CampaignCollectiveResult.js';
 import { buildCampaignLearningContent } from './build-campaign-learning-content.js';
 
 const buildCampaignCollectiveResult = function ({
