+++ api/tests/unit/application/campaign/index_test.js
@@ -4,33 +4,6 @@ import * as moduleUnderTest from '../../../../lib/application/campaigns/index.js
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | Router | campaign-router ', function () {
-  describe('GET /api/campaigns/{id}/collective-results', function () {
-    it('should return 200', async function () {
-      // given
-      sinon.stub(campaignController, 'getCollectiveResult').callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/1/collective-results');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return 400 with an invalid campaign id', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/campaigns/invalid/collective-results');
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
-
   describe('GET /api/campaigns/{id}/analyses', function () {
     it('should return 200', async function () {
       // given
